requirements.txt
future
h5py
matplotlib
numpy
pillow
plotly>=4.0.0
pyyaml
pandas
scikit-learn
scipy
torch>=1.6.0,<=1.9.0
torchvision>=0.7.0,<=0.10.0
kornia==0.5.0
wandb
tqdm
